Fetch user teensies server side on teensies/mine page
